add support building with commonlib
This requires common_lib to be built first and foremost as Makefile
won't check whether static library file exists, but it will quit as if
error occurs that such file is not there automatically. So we get this
for free from Makefile's target dependency itself.

main.cpp is refactored from extracted and refactored out functionality
(functions) from pcalc into commonlib.