usage text: use program_invocation_short_name as program name when print out help text
This has a real use case of having multiple instance of pcalc
but with different name. So it print out as part of help text
according to its real file name as exists on the file system.

Imagine having pcalc-bitkub, and another for pcalc-binance.
With different of fee between both binaries.